browser simulater: component architecture, widget / ui updates (#3063)
- Revert the `App` back to a generic component manager. It's best if
business logic can stay away from that as it defines how the core TUI
components interact.
- Use
[tui-textarea](https://github.com/rhysd/tui-textarea?tab=readme-ov-file#tui-textarea)
for editing. This is nicer as it has selection and history (not
connected yet) built-in.
- Use
[tui-logger](https://github.com/gin66/tui-logger?tab=readme-ov-file#tui-logger)
for the logging widget.

I think the core structure is good as is. Components can combine UI and
view models, the `App` and `Tui` drive the application, the `browser`
submodules deal with browser / hyper interaction.

---


![image](https://github.com/user-attachments/assets/75c628a3-6372-4e25-a62d-1049d7b0913c)


![image](https://github.com/user-attachments/assets/07e322e7-5de6-48f5-a2e7-3cef62c3f9b0)

---

I'm not happy yet about the term colors. I'm mostly using a dark
background in my terminal and the existing colors were not readable. I
changed things to look better with a dark bg


![image](https://github.com/user-attachments/assets/abfa6739-3946-484c-8850-d307cd9b4a76)

but with a light bg it is less readable now :/


![image](https://github.com/user-attachments/assets/0dd37a14-e203-45b4-ae9f-c1bb54c43af7)

We might need to force a single background color to avoid this. k9s and
btop do the same mostly.

---

Update: fix some of the coloring issues, still light bg is not ideal.

---

Depends on: https://github.com/hypervideo/hyper.video/pull/3062

---------

Co-authored-by: Tibor Hudik <tibor@hudik.eu>