browser simulator: fake audio and video from files and urls (#3068)
This adds support to specify arbitrary audio and video files. Files can
either be local paths or URLs which will be downloaded. The simulator
will take care of pre-processing those files to convert into yuv420p and
wav formats so that those can be used as "fake media" in
Chrome/Chromium.


![image](https://github.com/user-attachments/assets/e8ee1bf2-546c-435e-a3e1-74e24dad1b59)