improved browser simulator settings (#3318)
This addresses comments of
https://github.com/hypervideo/hyper.video/pull/3128#pullrequestreview-2916288585
and allows setting the various configuration options for participant
behavior before browser start. In the participants view noise
suppression and resolution can now be set with list inputs.

In particular, it cleans up the global api and uses it for most settings
functionality. Settings can be configured before a browser starts and
are persistent.