multi workspace (#3652)
Instead of a single cargo workspace, this introduces multiple workspaces
(basically per "app") to avoid dependency issues. cargo workspaces are
meant to bundle multiple crates for roughly the same build targets and
settings but our setup has evolved enough that it is not longer
practical to have only one workspace.

The following workspaces exist:
- backend
- frontend/core
- frontend/ios
- tools/audio-benchmarks
- tools/client-simulator

Other crates (mostly `shared/`) are standalone crates as they are pulled
in as libraries into the workspaces.