end to end tests with playwright (#3787)
This sets up end-to-end tests with playwright and removes the cypress
setup.

Tests live in `testing/e2e-web/` and require a hyper server to run in
the background (they will start one if there isn't one running on 8081).

This is a followup to #2494 that added the base setup for playwright.